refactor: 목표추가관련 로직 분리
